Fix symlink removal

+++ monty/tempfile.py
@@ -126,5 +126,5 @@ class ScratchDir:
 
             os.chdir(self.cwd)
             remove(self.tempdir)
-            if self.remove_symlink_on_exit:
+            if self.remove_symlink_on_exit and os.path.exists(ScratchDir.SCR_LINK):
                 os.remove(ScratchDir.SCR_LINK)
